- changed data class names to reflect Kotlin best practices - reformatted data class attributes for CleaningTask and MyRoom - Moved saveItem() and saveCleaningTask() to MainViewModel
