+++ app/src/main/java/com/example/mycleaninglog/ExpandableCard.kt
@@ -21,16 +21,13 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.TextUnit
 import androidx.compose.ui.unit.dp
-import com.example.mycleaninglog.dto.cleaningTask
-import com.example.mycleaninglog.dto.myRoom
+import com.example.mycleaninglog.dto.CleaningTask
+import com.example.mycleaninglog.dto.MyRoom
 import com.example.mycleaninglog.ui.theme.Gray
-import com.google.firebase.platforminfo.DefaultUserAgentPublisher
 
 
 @ExperimentalMaterialApi
@@ -42,8 +39,8 @@ fun ExpandableCard(
     title: String,
     titleFontSize: TextUnit = MaterialTheme.typography.h6.fontSize,
     titleFontWeight: FontWeight = FontWeight.Bold,
-    myRooms: List<myRoom> = ArrayList<myRoom>(),
-    myCleaningTasks: List<cleaningTask>,
+    myRooms: List<MyRoom> = ArrayList<MyRoom>(),
+    myCleaningTasks: List<CleaningTask>,
     viewModel: MainViewModel
     ) {
     //Used for Expandable card level 1 functionality
@@ -57,7 +54,7 @@ fun ExpandableCard(
     //val context = LocalContext.current
     //var selectedMyRoom: myRoom? = null
     //var viewModel = inViewModel
-viewModel.listenToMyRooms()
+    viewModel.listenToMyRooms()
 
     //building the card
     val dark = isSystemInDarkTheme()
@@ -131,7 +128,7 @@ viewModel.listenToMyRooms()
                         //bedroom
                         DropdownMenuItem(
                             onClick = {
-                                saveItem("Bedroom", "Bed", viewModel)
+                                viewModel.saveItem("Bedroom", "Bed", viewModel)
                                 addRoomShowMenu = !addRoomShowMenu
                             })
                         { Text(text = "Bedroom") }
@@ -139,7 +136,7 @@ viewModel.listenToMyRooms()
                         //bathroom
                         DropdownMenuItem(
                             onClick = {
-                                saveItem("Bathroom", "Bath", viewModel)
+                                viewModel.saveItem("Bathroom", "Bath", viewModel)
                                 addRoomShowMenu = !addRoomShowMenu
                             })
                         { Text(text = "Bathroom") }
@@ -148,7 +145,7 @@ viewModel.listenToMyRooms()
                         DropdownMenuItem(
                             onClick = {
                                 //addRoomList.add("Kitchen")
-                                saveItem("Kitchen", "Kit", viewModel)
+                                viewModel.saveItem("Kitchen", "Kit", viewModel)
                                 addRoomShowMenu = !addRoomShowMenu
                             })
                         { Text(text = "Kitchen") }
@@ -157,7 +154,7 @@ viewModel.listenToMyRooms()
                         DropdownMenuItem(
                             onClick = {
                                 //addRoomList.add("Dining Room")
-                                saveItem("Dining Room", "Din", viewModel)
+                                viewModel.saveItem("Dining Room", "Din", viewModel)
                                 addRoomShowMenu = !addRoomShowMenu
                             })
                         { Text(text = "Dining Room") }
@@ -166,7 +163,7 @@ viewModel.listenToMyRooms()
                         DropdownMenuItem(
                             onClick = {
                                 //addRoomList.add("Living Room")
-                                saveItem("Living Room", "Liv", viewModel)
+                                viewModel.saveItem("Living Room", "Liv", viewModel)
                                 addRoomShowMenu = !addRoomShowMenu
                             })
                         { Text(text = "Living Room") }
@@ -175,7 +172,7 @@ viewModel.listenToMyRooms()
                         DropdownMenuItem(
                             onClick = {
                                 //addRoomList.add("Outdoors")
-                                saveItem("Outdoors", "Out", viewModel)
+                                viewModel.saveItem("Outdoors", "Out", viewModel)
                                 addRoomShowMenu = !addRoomShowMenu
                             })
                         { Text(text = "Outdoors") }
@@ -184,7 +181,7 @@ viewModel.listenToMyRooms()
                         DropdownMenuItem(
                             onClick = {
                                 //addRoomList.add("Garage")
-                                saveItem("Garage", "Gar", viewModel)
+                                viewModel.saveItem("Garage", "Gar", viewModel)
                                 addRoomShowMenu = !addRoomShowMenu
                             })
                         { Text(text = "Garage") }
@@ -193,7 +190,7 @@ viewModel.listenToMyRooms()
                         DropdownMenuItem(
                             onClick = {
                                 //addRoomList.add("Utility Room")
-                                saveItem("Utility Room", "Util", viewModel)
+                                viewModel.saveItem("Utility Room", "Util", viewModel)
                                 addRoomShowMenu = !addRoomShowMenu
                             })
                         { Text(text = "Utility Room") }
@@ -229,93 +226,13 @@ viewModel.listenToMyRooms()
                     //following code needs to be replaced with a list of tasks in order from most needed to least
                     Text(text = "here are the upcoming tasks")
                 }
-
             }
         }
 
     }
 }
 
-fun saveItem(roomName: String, roomID: String,viewModel: MainViewModel ){
-    var preConRoom = myRoom().apply{
-        myRoomName = roomName
-        myRoomID = roomID
-    }
-    //creates prebuilt tasks into bedroom
-    viewModel.saveRoom(preConRoom)
-    if(preConRoom.myRoomName == "Bedroom"){
-        saveCleaningTask("Vacuum", "VAC", viewModel, preConRoom)
-        saveCleaningTask("Dusting", "DUS", viewModel, preConRoom)
-        saveCleaningTask("Wash Bedding", "WAS", viewModel, preConRoom)
-        saveCleaningTask("Laundry", "LAU", viewModel, preConRoom)
-    }
-    //creates prebuilt tasks into bathroom
-    if(preConRoom.myRoomName == "Bathroom"){
-        saveCleaningTask("Dusting", "DUS", viewModel, preConRoom)
-        saveCleaningTask("Scrub Floors", "FLO", viewModel, preConRoom)
-        saveCleaningTask("Scrub Vanity", "VAN", viewModel, preConRoom)
-        saveCleaningTask("Scrub Shower", "SHO", viewModel, preConRoom)
-        saveCleaningTask("Scrub Bath Tub", "TUB", viewModel, preConRoom)
-        saveCleaningTask("Scrub Toilet", "TOI", viewModel, preConRoom)
-        saveCleaningTask("Wipe Down Mirror", "MIR", viewModel, preConRoom)
-    }
-    //creates prebuilt tasks into kitchen
-    if(preConRoom.myRoomName == "Kitchen"){
-        saveCleaningTask("Dusting", "DUS", viewModel, preConRoom)
-        saveCleaningTask("Scrub Floors", "FLO", viewModel, preConRoom)
-        saveCleaningTask("Scrub Counter Tops", "COU", viewModel, preConRoom)
-        saveCleaningTask("Scrub Sink", "SIN", viewModel, preConRoom)
-        saveCleaningTask("Clean Microwave", "MIC", viewModel, preConRoom)
-        saveCleaningTask("Clean Fridge", "Fri", viewModel, preConRoom)
-        saveCleaningTask("Clean Oven", "OVE", viewModel, preConRoom)
-        saveCleaningTask("Clean Cabinets/Drawers", "CAB", viewModel, preConRoom)
-        saveCleaningTask("Wash Kitchen Towels/Rags", "TOW", viewModel, preConRoom)
-    }
-    //creates prebuilt tasks into dining room
-    if(preConRoom.myRoomName == "Dining Room"){
-        saveCleaningTask("Dusting", "DUS", viewModel, preConRoom)
-        saveCleaningTask("Scrub Floors", "FLO", viewModel, preConRoom)
-        saveCleaningTask("Clean Table", "TAB", viewModel, preConRoom)
-        saveCleaningTask("Wash Linens", "Lin", viewModel, preConRoom)
-    }
-    //creates prebuilt tasks into living room
-    if(preConRoom.myRoomName == "Living Room"){
-        saveCleaningTask("Dusting", "DUS", viewModel, preConRoom)
-        saveCleaningTask("Vacuum", "VAC", viewModel, preConRoom)
-        saveCleaningTask("Scrub Floors", "FLO", viewModel, preConRoom)
-        saveCleaningTask("Clean Upholstery", "UPH", viewModel, preConRoom)
-    }
-    //creates prebuilt tasks into outdoors
-    if(preConRoom.myRoomName == "Outdoors"){
-        saveCleaningTask("Mow", "MOW", viewModel, preConRoom)
-        saveCleaningTask("Pull Weeds", "WEE", viewModel, preConRoom)
-        saveCleaningTask("Water Flowers/Garden", "WAT", viewModel, preConRoom)
-        saveCleaningTask("Clean Grill", "GRI", viewModel, preConRoom)
-        saveCleaningTask("Clean Deck/Porch", "DEC", viewModel, preConRoom)
-        saveCleaningTask("Clean Windows", "WIN", viewModel, preConRoom)
-        saveCleaningTask("Clean Siding/Brick", "GRI", viewModel, preConRoom)
-        saveCleaningTask("Take Out Trash", "GRI", viewModel, preConRoom)
-    }
-    //creates prebuilt tasks into garage
-    if(preConRoom.myRoomName == "Garage"){
-        saveCleaningTask("Dusting", "DUS", viewModel, preConRoom)
-        saveCleaningTask("Clean Garage Door", "VAC", viewModel, preConRoom)
-        saveCleaningTask("Scrub Floors", "FLO", viewModel, preConRoom)
-    }
-    //creates prebuilt tasks into utility room
-    if(preConRoom.myRoomName == "Utility Room"){
-        saveCleaningTask("Dusting", "DUS", viewModel, preConRoom)
-        saveCleaningTask("Replace Furnace Filter", "VAC", viewModel, preConRoom)
-        saveCleaningTask("Scrub Floors", "FLO", viewModel, preConRoom)
-    }
-}
 
-//creates a cleaning task and sends to viewmodel where it saves tasks to a room in database
-fun saveCleaningTask(taskName: String, taskID: String, viewModel: MainViewModel, preConRoom: myRoom){
-    var preConTask = cleaningTask().apply{
-        cleaningTaskName = taskName
-        cleaningTaskId = taskID
-    }
-    viewModel.saveCleaningTask(preConTask, preConRoom)
-}
+
+
