+++ app/src/main/java/com/example/mycleaninglog/ExpandableCardLevelTwo.kt
@@ -1,10 +1,7 @@
 package com.example.mycleaninglog
 
 
-import android.app.AlertDialog
 import android.content.Context
-import android.widget.Button
-import android.widget.EditText
 import androidx.compose.animation.animateContentSize
 import androidx.compose.animation.core.LinearOutSlowInEasing
 import androidx.compose.animation.core.animateFloatAsState
@@ -14,11 +11,8 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.*
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.ArrowDropDown
 import androidx.compose.material.icons.filled.Menu
 import androidx.compose.runtime.*
@@ -27,28 +21,21 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.colorResource
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.TextUnit
 import androidx.compose.ui.unit.dp
-import androidx.fragment.app.FragmentActivity
-import androidx.fragment.app.FragmentManager
-import com.example.mycleaninglog.dto.cleaningTask
-import com.example.mycleaninglog.dto.myRoom
+import com.example.mycleaninglog.dto.CleaningTask
+import com.example.mycleaninglog.dto.MyRoom
 
 @ExperimentalMaterialApi
 @Composable
 fun ExpandableCardLevelTwo(
     //expandable card 2 variables
     c: Context,
-    selectedRoom: myRoom,
+    selectedRoom: MyRoom,
     viewModel: MainViewModel,
-    myCleaningTasks: List<cleaningTask>,
-    myRooms: List<myRoom>
+    myCleaningTasks: List<CleaningTask>,
+    myRooms: List<MyRoom>
     ) {
 
     // used for expandable card level 2 functionality
