+++ app/src/main/java/com/example/mycleaninglog/MainActivity.kt
@@ -13,10 +13,9 @@ import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.mycleaninglog.dto.myRoom
+import com.example.mycleaninglog.dto.MyRoom
 import com.example.mycleaninglog.dto.User
 import com.example.mycleaninglog.ui.theme.MyCleaningLogTheme
 import com.firebase.ui.auth.AuthUI
@@ -123,7 +122,7 @@ class MainActivity : ComponentActivity() {
 
 @ExperimentalMaterialApi
 @Composable
-fun HomeScreen(myRooms: List<myRoom> = ArrayList<myRoom>()) {
+fun HomeScreen(myRooms: List<MyRoom> = ArrayList<MyRoom>()) {
     //ExpandableCard(title = "Rooms", myRooms = myRooms, )
     //ExpandableCard(title = "Common Tasks")
     //ExpandableCard(title = "Upcoming Tasks")
