+++ app/src/main/java/com/example/mycleaninglog/RoomSettingsDialogBox.kt
@@ -1,17 +1,11 @@
 package com.example.mycleaninglog
 
 import android.content.Context
-import android.util.Log
-import android.widget.Toast
-import androidx.activity.ComponentActivity
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.AlertDialog
 import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.Text
-import androidx.compose.material.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -19,25 +13,21 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.colorResource
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.DialogProperties
 
 import androidx.compose.material.*
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Close
-import androidx.compose.material.icons.filled.Menu
 
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
-import com.example.mycleaninglog.dto.myRoom
+import com.example.mycleaninglog.dto.MyRoom
 
 
 @ExperimentalMaterialApi
 @Composable
-fun RoomSettingsDialogBox(c: Context, selectedRoomSettings: myRoom, viewModel: MainViewModel) {
+fun RoomSettingsDialogBox(c: Context, selectedRoomSettings: MyRoom, viewModel: MainViewModel) {
     val openDialog = remember { mutableStateOf(true) }
     val nameChange = remember {mutableStateOf("")}
     val addTask = remember {mutableStateOf("")}
