+++ app/src/main/java/com/example/mycleaninglog/dto/CleaningTask.kt
@@ -2,7 +2,11 @@ package com.example.mycleaninglog.dto
 
 import com.google.firebase.firestore.Exclude
 
-data class cleaningTask(var cleaningTaskName:String = "", var cleaningTaskId:String = "", var uniqueID:String = ""){
+data class CleaningTask(
+    var cleaningTaskName:String = "",
+    var cleaningTaskId:String = "",
+    var uniqueID:String = ""
+){
 
     override fun toString(): String {
         return "$cleaningTaskName"
