+++ app/src/main/java/com/example/mycleaninglog/dto/MyRoom.kt
@@ -2,7 +2,12 @@ package com.example.mycleaninglog.dto
 
 import com.google.firebase.firestore.Exclude
 
-data class myRoom(var myRoomName: String = "", var myRoomID:String = "", var uniqueID:String = "", var expanded: Boolean = false ){
+data class MyRoom(
+    var myRoomName: String = "",
+    var myRoomID:String = "",
+    var uniqueID:String = "",
+    var expanded: Boolean = false
+){
 
 
     override fun toString(): String {
